Only keep keccak in utils for now
