+++ addpeer.go
@@ -7,6 +7,8 @@ import (
 
 	inet "github.com/libp2p/go-libp2p-net"
 	pstore "github.com/libp2p/go-libp2p-peerstore"
+	peer "github.com/libp2p/go-libp2p-peer"
+	ma "github.com/multiformats/go-multiaddr"
 
 	pbmsg "github.com/ethresearch/sharding-p2p-poc/pb/message"
 )
@@ -19,6 +21,33 @@ const addPeerResponse = "/addPeer/response/0.0.1"
 type AddPeerProtocol struct {
 	node *Node     // local host
 	done chan bool // only for demo purposes to stop main from terminating
+}
+
+func parseAddr(addrString string) (peerID peer.ID, protocolAddr ma.Multiaddr) {
+	// The following code extracts target's the peer ID from the
+	// given multiaddress
+	ipfsaddr, err := ma.NewMultiaddr(addrString) // ipfsaddr=/ip4/127.0.0.1/tcp/10000/ipfs/QmVmDaabYcS3pn23KaFjkdw6hkReUUma8sBKqSDHrPYPd2
+	if err != nil {
+		log.Fatalln(err)
+	}
+
+	pid, err := ipfsaddr.ValueForProtocol(ma.P_IPFS) // pid=QmVmDaabYcS3pn23KaFjkdw6hkReUUma8sBKqSDHrPYPd2
+	if err != nil {
+		log.Fatalln(err)
+	}
+
+	peerid, err := peer.IDB58Decode(pid) // peerid=<peer.ID VmDaab>
+	if err != nil {
+		log.Fatalln(err)
+	}
+
+	// Decapsulate the /ipfs/<peerID> part from the target
+	// /ip4/<a.b.c.d>/ipfs/<peer> becomes /ip4/<a.b.c.d>
+	targetPeerAddr, _ := ma.NewMultiaddr(
+		fmt.Sprintf("/ipfs/%s", peer.IDB58Encode(peerid)),
+	)
+	targetAddr := ipfsaddr.Decapsulate(targetPeerAddr)
+	return peerid, targetAddr
 }
 
 func NewAddPeerProtocol(node *Node) *AddPeerProtocol {
